open a new file for hybrid checkpoint_io
